# src/main/java/io/swagger/api/AttendanceApi.java
/**
 * NOTE: This class is auto generated by the swagger code generator program (2.3.1).
 * https://github.com/swagger-api/swagger-codegen
 * Do not edit the class manually.
 */
package io.swagger.api;

import ru.fds.hrdepartment.dto.AttendanceSheetDto;
import io.swagger.annotations.*;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestHeader;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;
import org.springframework.web.bind.annotation.RequestParam;
import org.springframework.web.bind.annotation.RequestPart;
import org.springframework.web.multipart.MultipartFile;

import javax.validation.Valid;
import javax.validation.constraints.*;
import java.util.List;
@javax.annotation.Generated(value = "io.swagger.codegen.languages.SpringCodegen", date = "2020-09-11T10:14:57.918+03:00")

@Api(value = "attendance", description = "the attendance API")
public interface AttendanceApi {

    @ApiOperation(value = "get attendance by id", nickname = "getAttendance", notes = "", response = AttendanceSheetDto.class, tags={ "attendanceSheetController", })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "OK", response = AttendanceSheetDto.class),
        @ApiResponse(code = 401, message = "Unauthorized"),
        @ApiResponse(code = 403, message = "Forbidden"),
        @ApiResponse(code = 404, message = "Not Found") })
    @RequestMapping(value = "/attendance/{attendanceId}",
        produces = { "*/*" }, 
        method = RequestMethod.GET)
    ResponseEntity<AttendanceSheetDto> getAttendance(@ApiParam(value = "attendanceId",required=true) @PathVariable("attendanceId") Long attendanceId);


    @ApiOperation(value = "get attendance by employee id", nickname = "getAttendanceByEmployee", notes = "", response = AttendanceSheetDto.class, responseContainer = "List", tags={ "attendanceSheetController", })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "OK", response = AttendanceSheetDto.class, responseContainer = "List"),
        @ApiResponse(code = 401, message = "Unauthorized"),
        @ApiResponse(code = 403, message = "Forbidden"),
        @ApiResponse(code = 404, message = "Not Found") })
    @RequestMapping(value = "/attendance/emp/{employeeId}",
        produces = { "*/*" }, 
        method = RequestMethod.GET)
    ResponseEntity<List<AttendanceSheetDto>> getAttendanceByEmployee(@ApiParam(value = "employeeId",required=true) @PathVariable("employeeId") Long employeeId);


    @ApiOperation(value = "insert new attendance", nickname = "insertAttendance", notes = "", response = AttendanceSheetDto.class, tags={ "attendanceSheetController", })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "OK", response = AttendanceSheetDto.class),
        @ApiResponse(code = 201, message = "Created"),
        @ApiResponse(code = 401, message = "Unauthorized"),
        @ApiResponse(code = 403, message = "Forbidden"),
        @ApiResponse(code = 404, message = "Not Found") })
    @RequestMapping(value = "/attendance/new",
        produces = { "*/*" }, 
        consumes = { "application/json" },
        method = RequestMethod.POST)
    ResponseEntity<AttendanceSheetDto> insertAttendance(@ApiParam(value = "attendanceSheetDto" ,required=true )  @Valid @RequestBody AttendanceSheetDto attendanceSheetDto);


    @ApiOperation(value = "update attendance", nickname = "updateAttendance", notes = "", response = AttendanceSheetDto.class, tags={ "attendanceSheetController", })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "OK", response = AttendanceSheetDto.class),
        @ApiResponse(code = 201, message = "Created"),
        @ApiResponse(code = 401, message = "Unauthorized"),
        @ApiResponse(code = 403, message = "Forbidden"),
        @ApiResponse(code = 404, message = "Not Found") })
    @RequestMapping(value = "/attendance/update",
        produces = { "*/*" }, 
        consumes = { "application/json" },
        method = RequestMethod.PUT)
    ResponseEntity<AttendanceSheetDto> updateAttendance(@ApiParam(value = "attendanceSheetDto" ,required=true )  @Valid @RequestBody AttendanceSheetDto attendanceSheetDto);

}