# src/main/resources/api.yaml
swagger: '2.0'
info:
  description: Api Documentation
  version: '1.0'
  title: Api Documentation
  termsOfService: 'urn:tos'
  contact: {}
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0'
host: 'localhost:8080'
basePath: /
...
paths:
  /attendance/new:
    post:
      tags:
        - attendanceSheetController
      summary: insert new attendance
      operationId: insertAttendance
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - in: body
          name: attendanceSheetDto
          description: attendanceSheetDto
          required: true
          schema:
            $ref: '#/definitions/AttendanceSheetDto'
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/AttendanceSheetDto'
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      deprecated: false
  /attendance/update:
    put:
      tags:
        - attendanceSheetController
      summary: update attendance
      operationId: updateAttendance
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - in: body
          name: attendanceSheetDto
          description: attendanceSheetDto
          required: true
          schema:
            $ref: '#/definitions/AttendanceSheetDto'
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/AttendanceSheetDto'
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      deprecated: false
  '/attendance/emp/{employeeId}':
    get:
      tags:
        - attendanceSheetController
      summary: get attendance by employee id
      operationId: getAttendanceByEmployee
      produces:
        - '*/*'
      parameters:
        - name: employeeId
          in: path
          description: employeeId
          required: true
          type: integer
          format: int64
      responses:
        '200':
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/AttendanceSheetDto'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      deprecated: false
  '/attendance/{attendanceId}':
    get:
      tags:
        - attendanceSheetController
      summary: get attendance by id
      operationId: getAttendance
      produces:
        - '*/*'
      parameters:
        - name: attendanceId
          in: path
          description: attendanceId
          required: true
          type: integer
          format: int64
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/AttendanceSheetDto'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      deprecated: false
  /dep/new:
    post:
      tags:
        - departmentController
      summary: insert new department
      operationId: insertDepartment
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - in: body
          name: departmentDto
          description: departmentDto
          required: true
          schema:
            $ref: '#/definitions/DepartmentDto'
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/DepartmentDto'
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      deprecated: false
  /dep/total_hour:
    get:
      tags:
        - departmentController
      summary: get total work hours in department today
      operationId: getWorkHoursInDepartment
      produces:
        - '*/*'
      parameters:
        - name: departmentId
          in: query
          description: departmentId
          required: true
          type: integer
          format: int64
      responses:
        '200':
          description: OK
          schema:
            type: integer
            format: int32
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      deprecated: false
  /dep/update:
    put:
      tags:
        - departmentController
      summary: update department
      operationId: updateDepartment
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - in: body
          name: departmentDto
          description: departmentDto
          required: true
          schema:
            $ref: '#/definitions/DepartmentDto'
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/DepartmentDto'
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      deprecated: false
...
definitions:
  AttendanceSheetDto:
    type: object
    properties:
      employeeId:
        type: integer
        format: int64
      hourAtWork:
        type: integer
        format: int32
      id:
        type: integer
        format: int64
      typeOfAttendance:
        type: string
        enum:
          - AT_WORK
          - ABSENCE
          - ON_SICK_LEAVE
          - ON_VACATION
      workDate:
        type: string
        format: date
    title: AttendanceSheetDto
  DepartmentDto:
    type: object
    properties:
      id:
        type: integer
        format: int64
      name:
        type: string
    title: DepartmentDto
...